src/training_pipeline/finetuning/requirements.txt
numpy==1.24.0
pandas==2.2.2
peft==0.11.0
datasets==2.19.1
transformers==4.40.2
safetensors==0.4.3
comet_ml==3.42.0
accelerate==0.30.1
bitsandbytes==0.42.0
pydantic_settings==2.2.1
scikit-learn==1.4.2
qwak-sdk==0.5.85
structlog==24.1.0
